refactor(Reservation) : reservationDAO의 예외처리 service 계층으로 이동(한글로 메세지 사용)

+++ src/main/java/roomescape/service/ReservationService.java
@@ -2,9 +2,13 @@ package roomescape.service;
 
 
 import java.util.List;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 import roomescape.dao.reservation.ReservationJdbcDAO;
 import roomescape.entity.Reservation;
+import roomescape.exception.DataInvalidException;
+import roomescape.exception.InvalidException;
+import roomescape.exception.ReservationNotFoundException;
 
 @Service
 public class ReservationService {
@@ -16,18 +20,34 @@ public class ReservationService {
     }
 
     public List<Reservation> getAllReservations() {
-        return reservationJdbcDAO.getAll();
+        try {
+            return reservationJdbcDAO.getAll();
+        } catch (DataAccessException e) {
+            throw new DataInvalidException("예약 데이터를 불러오는데 실패했습니다.");
+        }
     }
 
     public Reservation createReservation(Reservation reservation) {
-        return reservationJdbcDAO.create(reservation);
+        try {
+            return reservationJdbcDAO.create(reservation);
+        } catch (DataAccessException e) {
+            throw new DataInvalidException("예약을 생성하는데 실패했습니다.");
+        }
     }
 
     public Reservation getReservationById(long id) {
-        return reservationJdbcDAO.getById(id);
+        Reservation reservation = reservationJdbcDAO.getById(id);
+        if (reservation == null) {
+            throw new ReservationNotFoundException("예약을 찾을 수 없습니다." + id);
+        }
+        return reservation;
     }
 
     public void deleteReservationById(long id) {
+        Reservation reservation = reservationJdbcDAO.getById(id);
+        if (reservation == null) {
+            throw new InvalidException("예약을 찾을 수 없습니다." + id);
+        }
         reservationJdbcDAO.delete(id);
     }
